Ignore EOF while reading parameters
