+++ pkg/kapis/devops/v1alpha4/handler.go
@@ -2,10 +2,10 @@ package v1alpha4
 
 import (
 	"context"
-	"errors"
+	"io"
+	"strconv"
+
 	"github.com/emicklei/go-restful"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"kubesphere.io/devops/pkg/api"
 	"kubesphere.io/devops/pkg/api/devops/v1alpha3"
 	"kubesphere.io/devops/pkg/api/devops/v1alpha4"
@@ -13,7 +13,6 @@ import (
 	"kubesphere.io/devops/pkg/client/devops"
 	resourcesV1alpha3 "kubesphere.io/devops/pkg/models/resources/v1alpha3"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"strconv"
 )
 
 type apiHandlerOption struct {
@@ -82,7 +81,7 @@ func (h *apiHandler) createPipelineRuns() restful.RouteFunction {
 		pipName := request.PathParameter("pipeline")
 		branch := request.QueryParameter("branch")
 		payload := devops.RunPayload{}
-		if err := request.ReadEntity(&payload); err != nil {
+		if err := request.ReadEntity(&payload); err != nil && err != io.EOF {
 			api.HandleBadRequest(response, request, err)
 			return
 		}
@@ -110,46 +109,5 @@ func (h *apiHandler) createPipelineRuns() restful.RouteFunction {
 		}
 
 		_ = response.WriteEntity(pr)
-	}
-}
-
-func getScm(ps *v1alpha3.PipelineSpec, branch string) (*v1alpha4.SCM, error) {
-	var scm *v1alpha4.SCM
-	if ps.Type == v1alpha3.MultiBranchPipelineType {
-		if branch == "" {
-			return nil, errors.New("missing branch name for running a multi-branch Pipeline")
-		}
-		// TODO validate if the branch dose exist
-		// we can not determine what is reference type here. So we set reference name only for now
-		scm = &v1alpha4.SCM{
-			RefName: branch,
-			RefType: "",
-		}
-	}
-	return scm, nil
-}
-
-func getPipelineRef(pipeline *v1alpha3.Pipeline) *corev1.ObjectReference {
-	return &corev1.ObjectReference{
-		Kind:      pipeline.Kind,
-		Name:      pipeline.GetName(),
-		Namespace: pipeline.GetNamespace(),
-	}
-}
-
-func createPipelineRun(pipeline *v1alpha3.Pipeline, payload *devops.RunPayload, scm *v1alpha4.SCM) *v1alpha4.PipelineRun {
-	controllerRef := metav1.NewControllerRef(pipeline, pipeline.GroupVersionKind())
-	return &v1alpha4.PipelineRun{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName:    pipeline.GetName() + "-runs-",
-			Namespace:       pipeline.GetNamespace(),
-			OwnerReferences: []metav1.OwnerReference{*controllerRef},
-		},
-		Spec: v1alpha4.PipelineRunSpec{
-			PipelineRef:  getPipelineRef(pipeline),
-			PipelineSpec: &pipeline.Spec,
-			Parameters:   convertParameters(payload),
-			SCM:          scm,
-		},
 	}
 }
