+++ pkg/kapis/devops/v1alpha4/util.go
@@ -1,10 +1,15 @@
 package v1alpha4
 
 import (
+	"errors"
 	"fmt"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/selection"
+	"kubesphere.io/devops/pkg/api/devops/v1alpha3"
 	"kubesphere.io/devops/pkg/api/devops/v1alpha4"
 	"kubesphere.io/devops/pkg/apiserver/query"
 	"kubesphere.io/devops/pkg/client/devops"
@@ -52,4 +57,45 @@ func convertParameters(payload *devops.RunPayload) []v1alpha4.Parameter {
 		})
 	}
 	return parameters
+}
+
+func getScm(ps *v1alpha3.PipelineSpec, branch string) (*v1alpha4.SCM, error) {
+	var scm *v1alpha4.SCM
+	if ps.Type == v1alpha3.MultiBranchPipelineType {
+		if branch == "" {
+			return nil, errors.New("missing branch name for running a multi-branch Pipeline")
+		}
+		// TODO validate if the branch dose exist
+		// we can not determine what is reference type here. So we set reference name only for now
+		scm = &v1alpha4.SCM{
+			RefName: branch,
+			RefType: "",
+		}
+	}
+	return scm, nil
+}
+
+func getPipelineRef(pipeline *v1alpha3.Pipeline) *corev1.ObjectReference {
+	return &corev1.ObjectReference{
+		Kind:      pipeline.Kind,
+		Name:      pipeline.GetName(),
+		Namespace: pipeline.GetNamespace(),
+	}
+}
+
+func createPipelineRun(pipeline *v1alpha3.Pipeline, payload *devops.RunPayload, scm *v1alpha4.SCM) *v1alpha4.PipelineRun {
+	controllerRef := metav1.NewControllerRef(pipeline, pipeline.GroupVersionKind())
+	return &v1alpha4.PipelineRun{
+		ObjectMeta: metav1.ObjectMeta{
+			GenerateName:    pipeline.GetName() + "-runs-",
+			Namespace:       pipeline.GetNamespace(),
+			OwnerReferences: []metav1.OwnerReference{*controllerRef},
+		},
+		Spec: v1alpha4.PipelineRunSpec{
+			PipelineRef:  getPipelineRef(pipeline),
+			PipelineSpec: &pipeline.Spec,
+			Parameters:   convertParameters(payload),
+			SCM:          scm,
+		},
+	}
 }
